MAINT,API,BUG: Simplify `arr.fill`, fix it for some user dtypes
This modifies the logic to match that of `arr1d[0] = obj`, rather
than `arr1d[0] = np.asarray(obj)` (or similar).
That change could be considered good, but is unfortunately a slight
change in behaviour in some cases.

One (probably the main) change is that we allow casting ints to
unstructured voids, but we do not allow filling an unstructured void
with Python integers.

+++ numpy/core/tests/test_multiarray.py
@@ -67,7 +67,7 @@ def _aligned_zeros(shape, dtype=float, order="C", align=None):
     # data pointer --- so we use and allocate size+1
     buf = buf[offset:offset+size+1][:-1]
     data = np.ndarray(shape, dtype, buf, order=order)
-    data.fill(0)
+    data.fill(np.uint8(0))
     return data
